[AN] refactor(#139): 패키지 정리 및 register 관련 dto 수정
