+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/login/LoginPageResponse.kt
@@ -1,4 +1,4 @@
-package com.takseha.data.dto
+package com.takseha.data.dto.login.login
 
 
 import com.google.gson.annotations.SerializedName
