+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/login/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.takseha.data.dto
+package com.takseha.data.dto.login.login
 
 data class LoginRequest(
     val code: String,
