+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/login/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.takseha.data.dto
+package com.takseha.data.dto.login.login
 
 
 import com.google.gson.annotations.SerializedName
@@ -9,10 +9,10 @@ data class LoginResponse(
     @SerializedName("res_msg")
     val resMsg: String,
     @SerializedName("res_obj")
-    val tokenInfo: TokenInfo
+    val tokenInfo: LoginTokenInfo
 )
 
-data class TokenInfo(
+data class LoginTokenInfo(
     @SerializedName("access_token")
     val accessToken: String,
     @SerializedName("refresh_token")
