+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/login/Role.kt
@@ -1,4 +1,4 @@
-package com.takseha.data.dto
+package com.takseha.data.dto.login.login
 
 enum class Role {
     ADMIN, UNAUTH, USER, WITHDRAW
