+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/register/RegisterRequest.kt
@@ -1,0 +1,11 @@
+package com.takseha.data.dto.login.register
+
+
+import com.google.gson.annotations.SerializedName
+
+data class RegisterRequest(
+    @SerializedName("githubId")
+    val githubId: String,
+    @SerializedName("name")
+    val name: String
+)
