+++ android/gitudy/data/src/main/java/com/takseha/data/dto/login/register/RegisterResponse.kt
@@ -1,0 +1,20 @@
+package com.takseha.data.dto.login.register
+
+
+import com.google.gson.annotations.SerializedName
+
+data class RegisterResponse(
+    @SerializedName("res_code")
+    val resCode: Int,
+    @SerializedName("res_msg")
+    val resMsg: String,
+    @SerializedName("res_obj")
+    val tokenInfo: RegisterTokenInfo
+)
+
+data class RegisterTokenInfo(
+    @SerializedName("access_token")
+    val accessToken: String,
+    @SerializedName("refresh_token")
+    val refreshToken: String
+)
